+++ src/HueModuleInstance.cpp
@@ -1,4 +1,0 @@
-#include "HueModule.h"
-
-// Globale Instanz für OpenKNX Integration
-HueModule openknxHueModule;